printErrorsの共通化
